SurfaceFlinger: skip composition when presenting a single auto-refresh layer
When only one auto-refresh layer is presented on the primary display,
and is handled by the hardware composer, we can skip the composition
pass (invalidation and refresh) altogether. This reduces
SurfaceFlinger's CPU usage in a typical VR use case.

Together with setConsumerCanWait and the fix to keep auto-refresh state
in sync betwen producers and consumers, these provide the platform
building blocks necessary to implement a low-latency VR compositor using
an appropriately-configured fullscreen surface in an arbitrary app.

Test: manual - on Meta Quest 2, launch an arbitrary VR app and check
SurfaceFlinger's CPU usage.  Prior to this change, SurfaceFlinger spent
around 1.2ms of CPU time every frame iterating over layers and executing
HWC commits.  Also run screenrecord to verify it does not generate an
empty file. On bluejay, verify device boots and apps are visible and
interactable.

Change-Id: I7b21a99a8d157af2a197b97755687e000353ddb7