Memtrack: Fallback to HIDL implementation
Avoid attemptimg to get the HIDL service if we already got the AIDL
service. This can help to reduce the warning logs from servicemanager on
devices with an AIDL memtrack (this should be true for devices launching
on recent desserts).

servicemanager: Could not find android.hardware.memtrack.IMemtrack/default in the VINTF manifest.

On devices still using a HILD implementation
still expect to see service manager logs for the missing AIDL
implemetation.

Bug: 188388088
Test: Logcat
Change-Id: Ifd740ff2d0a17b6fc46b217e0ca8861ef7dc5f1e
Signed-off-by: Kalesh Singh <kaleshsingh@google.com>
Signed-off-by: minaripenguin <minaripenguin@users.noreply.github.com>